Moved status code hash map into own file
